requirements.txt
openai==0.28
streamlit==1.36.0
python-dotenv
black